Fix the pilot study script
I removed a module import that does not exist anymore.

I fixed the frame processing times array in the run() function. The
code produced an index-out-of-bounds error.

I fixed printing the progress bar. It was not printing all they way to
the end. Now the bar fills up when the sequence is done.