Fix notification bugs for GOT-10k experiments
I removed the SlackDropper class. Instead, I added the
ConsoleReporter to the GOT-10k exerpiments module. The new
reporter prints messages to the console.

I moved the make_slack_reporter() function from the Slack
reporter module to the GOT-10k experiments module, and
renamed it to make_notifier().

I fixed a bug in make_notifier(). If the Slack configuration does
not exist, the old code raised an exception. The new code returns
a ConsoleReporter() object.